Add skelethon environment setup
The skelethon is the base of the project with the first test scenario
passing providing to the up-coming development a stable environment
containing the base framework and test setup.

To prove the setup is correctly setup the 'not found' case of the
endpoint GET /persons/{id} was implemented and tested.